+++ CHANGELOG.md
@@ -495,6 +495,7 @@
 * [X] 实现Renderable
 * [X] 不做任何处理的`pass()`方法
 * [X] `coalesce()`和`coalesceOrNull()`
+* [X] `CharMatcher`和`StringMatcher`
 * [ ] 并发迭代的集合（ParallelCollection）
 * [ ] 多级列表和树的上钻下钻操作（drillUp drillDown）